requirements.txt
openai-whisper
sounddevice
numpy
scipy
openai
pydub
python-dotenv
pydantic
pydantic-settings
